Microservices/debtservice/requirements.txt
Flask
Flask-Cors
Flask-JWT-Extended
SQLAlchemy
python-dotenv
pyodbc
flask_sqlalchemy
flask_bcrypt